chore: move dbus interface to its own package
